refactor(ui): Rename Heading component to Section
Rename Heading component file to better reflect its purpose as a section
separator with title.
