+++ app/(playground)/p/[agentId]/beta-proto/left-menu/components/section.tsx
