+++ app/(playground)/p/[agentId]/beta-proto/left-menu/github-integration/github-integration.tsx
@@ -6,7 +6,7 @@ import {
 	SelectTrigger,
 	SelectValue,
 } from "../../components/select";
-import { Heading } from "../components/heading";
+import { Heading } from "../components/section";
 
 const mockRepositories = [
 	{
